create ui interface dashboard 3
