+++ app/build.gradle
@@ -16,7 +16,6 @@ android {
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
-
     buildTypes {
         release {
             minifyEnabled false
@@ -61,6 +60,20 @@ dependencies {
     // Recycler View
     implementation 'androidx.recyclerview:recyclerview:1.1.0'
 
+    //LifeCycle
+    implementation 'androidx.lifecycle:lifecycle-common:2.2.0'
+    implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.2.0'
+    implementation 'android.arch.lifecycle:extensions:2.2.0'
+    implementation 'androidx.lifecycle:lifecycle-livedata-ktx:2.2.0'
+
+    //Retrofit
+    implementation 'com.squareup.retrofit2:retrofit:2.9.0'
+    implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
+    implementation("com.squareup.okhttp3:logging-interceptor:4.8.1")
+
+    //Coroutines
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.4.1'
+
     // Tensor flow Lite
     implementation 'org.tensorflow:tensorflow-lite:2.2.0'
 
@@ -69,6 +82,10 @@ dependencies {
 
     // Coil
     implementation 'io.coil-kt:coil:1.1.0'
+
+//    // Here Maps
+//    implementation fileTree(dir: 'libs', include: ['*.jar', '*.aar'])
+//    implementation(name:'HERE-sdk', ext:'aar')
 
     // Testing
     testImplementation 'junit:junit:4.13.1'
