+++ app/src/main/java/com/goggxi/covid19detector/ui/MainActivity.kt
@@ -1,0 +1,12 @@
+package com.goggxi.covid19detector.ui
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.goggxi.covid19detector.R
+
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+    }
+}
