+++ app/src/main/java/com/goggxi/covid19detector/ui/informasi/InformasiFragment.kt
@@ -1,0 +1,40 @@
+package com.goggxi.covid19detector.ui.informasi
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.navigation.fragment.NavHostFragment
+import androidx.navigation.ui.setupWithNavController
+import com.goggxi.covid19detector.R
+import com.goggxi.covid19detector.databinding.FragmentInformasiBinding
+
+class InformasiFragment : Fragment() {
+    private lateinit var binding: FragmentInformasiBinding
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_informasi, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        binding = FragmentInformasiBinding.inflate(layoutInflater)
+        activity?.setContentView(binding.root)
+
+        // Nav host fragment
+        val host: NavHostFragment = activity?.supportFragmentManager
+            ?.findFragmentById(R.id.container) as NavHostFragment?
+            ?: return
+        // nav controller
+        val navController = host.navController
+
+        // Setup bottom navigation view
+        binding.bottomNavigationId.setupWithNavController(navController)
+
+    }
+}
