+++ app/src/main/res/drawable/bg_bottom_navigation.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape android:shape="rectangle" xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <solid android:color="#fff"/>
+    <corners android:topRightRadius="16dp"
+        android:topLeftRadius="16dp"/>
+</shape>
