+++ app/src/main/res/drawable/ic_add_foto.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="20dp"
+    android:height="20dp"
+    android:viewportWidth="20"
+    android:viewportHeight="20">
+  <path
+      android:pathData="M16.8421,5.2632V8.4105C16.8421,8.4105 14.7474,8.4211 14.7368,8.4105V5.2632H11.5789C11.5789,5.2632 11.5895,3.1684 11.5789,3.1579H14.7368V0H16.8421V3.1579H20V5.2632H16.8421ZM13.6842,9.4737V6.3158H10.5263V3.1579H2.1053C0.9474,3.1579 0,4.1053 0,5.2632V17.8947C0,19.0526 0.9474,20 2.1053,20H14.7368C15.8947,20 16.8421,19.0526 16.8421,17.8947V9.4737H13.6842ZM2.1053,17.8947L5.2632,13.6842L7.3684,16.8421L10.5263,12.6316L14.7368,17.8947H2.1053Z"
+      android:fillColor="#1B68B1"/>
+</vector>
