+++ app/src/main/res/drawable/ic_dashboaard.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="20dp"
+    android:height="20dp"
+    android:viewportWidth="20"
+    android:viewportHeight="20">
+  <path
+      android:pathData="M8.9863,10.5059H17.9726C17.9733,12.2834 17.4469,14.0212 16.4599,15.4994C15.4729,16.9777 14.0697,18.1301 12.4277,18.8108C10.7858,19.4915 8.9788,19.67 7.2354,19.3237C5.492,18.9773 3.8904,18.1218 2.6333,16.8651C1.3762,15.6085 0.52,14.0073 0.1729,12.2641C-0.1741,10.5208 0.0037,8.7138 0.6837,7.0715C1.3637,5.4293 2.5155,4.0256 3.9934,3.038C5.4713,2.0504 7.2088,1.5233 8.9863,1.5233V10.5059Z"
+      android:fillColor="#686868"/>
+  <path
+      android:pathData="M11.0137,8.9826H20C19.999,6.6 19.0518,4.3152 17.3667,2.6307C15.6815,0.9463 13.3964,-0 11.0137,0V8.9826Z"
+      android:fillColor="#686868"/>
+</vector>
