+++ app/src/main/res/drawable/ic_informasi.xml
@@ -1,0 +1,24 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="17dp"
+    android:height="20dp"
+    android:viewportWidth="17"
+    android:viewportHeight="20">
+  <path
+      android:pathData="M16.4504,16.9337H16.4467H14.4292V18.9402V18.9549H14.4587L16.4504,16.9669V16.9337Z"
+      android:fillColor="#686868"/>
+  <path
+      android:pathData="M15.5394,0H1.4754C1.0968,0.0162 0.7373,0.1709 0.4653,0.4347C0.1933,0.6986 0.0277,1.0532 0,1.4311V1.977V3.9466H1.0217C1.1401,3.6751 1.3496,3.4533 1.614,3.3197C1.8784,3.186 2.1811,3.1488 2.47,3.2144C2.7589,3.2801 3.0159,3.4444 3.1966,3.6792C3.3772,3.914 3.4704,4.2045 3.4598,4.5005C3.4493,4.7966 3.3358,5.0797 3.1389,5.3011C2.942,5.5224 2.674,5.6681 2.3811,5.7131C2.0883,5.758 1.789,5.6994 1.5347,5.5473C1.2805,5.3952 1.0873,5.1591 0.9885,4.8798H0V7.2699H1.0291C1.1504,6.9995 1.3624,6.7799 1.6284,6.6491C1.8944,6.5183 2.1977,6.4844 2.486,6.5534C2.7743,6.6223 3.0294,6.7897 3.2074,7.0267C3.3855,7.2637 3.4752,7.5554 3.4611,7.8515C3.447,8.1476 3.33,8.4295 3.1302,8.6485C2.9305,8.8675 2.6606,9.0099 2.367,9.0511C2.0735,9.0924 1.7748,9.0298 1.5224,8.8743C1.2701,8.7188 1.0799,8.4801 0.9848,8.1994H0V10.7002H0.9922C1.0928,10.4234 1.2867,10.1904 1.5407,10.0413C1.7946,9.8922 2.0926,9.8363 2.3833,9.8833C2.674,9.9304 2.9392,10.0773 3.1331,10.2989C3.3271,10.5204 3.4377,10.8027 3.4459,11.0971C3.4541,11.3915 3.3593,11.6795 3.1779,11.9114C2.9966,12.1434 2.7399,12.3049 2.4523,12.368C2.1647,12.4311 1.864,12.3919 1.6022,12.2571C1.3404,12.1223 1.1338,11.9004 1.018,11.6296H0V13.9829H1.0143C1.128,13.7101 1.3333,13.4856 1.5948,13.3481C1.8563,13.2105 2.1576,13.1686 2.4467,13.2294C2.7359,13.2902 2.9947,13.4501 3.1786,13.6813C3.3625,13.9126 3.4599,14.2008 3.4541,14.4962C3.4482,14.7916 3.3394,15.0757 3.1465,15.2995C2.9536,15.5233 2.6886,15.6727 2.3973,15.722C2.106,15.7714 1.8066,15.7175 1.5507,15.5697C1.2949,15.4218 1.0986,15.1894 0.9959,14.9123H0V17.8373V18.2578C0,18.9954 0.7377,19.4897 1.4754,19.4897H13.4812L13.5107,15.9267H17V1.4422C16.9809,1.0627 16.8204,0.7042 16.55,0.4372C16.2796,0.1703 15.9191,0.0143 15.5394,0ZM14.6283,13.6657H4.795C4.6658,13.6657 4.542,13.6144 4.4507,13.5231C4.3594,13.4318 4.3081,13.3079 4.3081,13.1788C4.3081,13.0497 4.3594,12.9258 4.4507,12.8345C4.542,12.7432 4.6658,12.6919 4.795,12.6919H14.6357C14.7648,12.6919 14.8887,12.7432 14.98,12.8345C15.0713,12.9258 15.1226,13.0497 15.1226,13.1788C15.1226,13.3079 15.0713,13.4318 14.98,13.5231C14.8887,13.6144 14.7648,13.6657 14.6357,13.6657H14.6283ZM14.6283,11.4969H4.795C4.6658,11.4969 4.542,11.4456 4.4507,11.3543C4.3594,11.2629 4.3081,11.1391 4.3081,11.01C4.3081,10.8809 4.3594,10.757 4.4507,10.6657C4.542,10.5744 4.6658,10.5231 4.795,10.5231H14.6357C14.7648,10.5231 14.8887,10.5744 14.98,10.6657C15.0713,10.757 15.1226,10.8809 15.1226,11.01C15.1226,11.1391 15.0713,11.2629 14.98,11.3543C14.8887,11.4456 14.7648,11.4969 14.6357,11.4969H14.6283ZM14.6283,9.3723H4.795C4.6658,9.3723 4.542,9.321 4.4507,9.2297C4.3594,9.1384 4.3081,9.0146 4.3081,8.8854C4.3081,8.7563 4.3594,8.6325 4.4507,8.5412C4.542,8.4499 4.6658,8.3986 4.795,8.3986H14.6357C14.7648,8.3986 14.8887,8.4499 14.98,8.5412C15.0713,8.6325 15.1226,8.7563 15.1226,8.8854C15.1226,9.0146 15.0713,9.1384 14.98,9.2297C14.8887,9.321 14.7648,9.3723 14.6357,9.3723H14.6283ZM14.6283,7.2478H4.795C4.6658,7.2478 4.542,7.1965 4.4507,7.1052C4.3594,7.0139 4.3081,6.89 4.3081,6.7609C4.3081,6.6318 4.3594,6.5079 4.4507,6.4166C4.542,6.3253 4.6658,6.274 4.795,6.274H14.6357C14.7648,6.274 14.8887,6.3253 14.98,6.4166C15.0713,6.5079 15.1226,6.6318 15.1226,6.7609C15.1226,6.89 15.0713,7.0139 14.98,7.1052C14.8887,7.1965 14.7648,7.2478 14.6357,7.2478H14.6283ZM14.6283,5.1675H4.795C4.6658,5.1675 4.542,5.1162 4.4507,5.0249C4.3594,4.9336 4.3081,4.8098 4.3081,4.6806C4.3081,4.5515 4.3594,4.4277 4.4507,4.3364C4.542,4.245 4.6658,4.1937 4.795,4.1937H14.6357C14.7648,4.1937 14.8887,4.245 14.98,4.3364C15.0713,4.4277 15.1226,4.5515 15.1226,4.6806C15.1226,4.8098 15.0713,4.9336 14.98,5.0249C14.8887,5.1162 14.7648,5.1675 14.6357,5.1675H14.6283Z"
+      android:fillColor="#686868"/>
+  <path
+      android:pathData="M2.7774,4.4593C2.7774,4.3368 2.7411,4.217 2.673,4.115C2.6049,4.0132 2.5081,3.9337 2.3949,3.8868C2.2817,3.8399 2.1571,3.8277 2.0369,3.8516C1.9167,3.8755 1.8062,3.9345 1.7196,4.0212C1.6329,4.1078 1.5739,4.2182 1.55,4.3384C1.5261,4.4586 1.5384,4.5832 1.5853,4.6965C1.6322,4.8097 1.7116,4.9065 1.8135,4.9745C1.9154,5.0426 2.0352,5.079 2.1577,5.079C2.3218,5.078 2.4788,5.0124 2.5948,4.8964C2.7108,4.7804 2.7764,4.6234 2.7774,4.4593ZM1.8258,4.4593C1.8258,4.3937 1.8453,4.3295 1.8817,4.2749C1.9182,4.2203 1.97,4.1778 2.0307,4.1526C2.0914,4.1275 2.1581,4.1209 2.2225,4.1337C2.2869,4.1466 2.3461,4.1782 2.3925,4.2246C2.4389,4.271 2.4705,4.3302 2.4833,4.3946C2.4961,4.459 2.4896,4.5257 2.4644,4.5863C2.4393,4.647 2.3968,4.6989 2.3422,4.7353C2.2876,4.7718 2.2234,4.7913 2.1577,4.7913C2.0707,4.7856 1.9892,4.7462 1.9307,4.6815C1.8722,4.6167 1.8412,4.5318 1.8442,4.4446L1.8258,4.4593Z"
+      android:fillColor="#686868"/>
+  <path
+      android:pathData="M2.7774,7.7789C2.7774,7.6563 2.7411,7.5365 2.673,7.4346C2.6049,7.3327 2.5081,7.2533 2.3949,7.2064C2.2817,7.1595 2.1571,7.1472 2.0369,7.1712C1.9167,7.1951 1.8062,7.2541 1.7196,7.3407C1.6329,7.4274 1.5739,7.5378 1.55,7.658C1.5261,7.7782 1.5384,7.9028 1.5853,8.016C1.6322,8.1293 1.7116,8.226 1.8135,8.2941C1.9154,8.3622 2.0352,8.3986 2.1577,8.3986C2.3221,8.3986 2.4797,8.3333 2.5959,8.2171C2.7121,8.1009 2.7774,7.9432 2.7774,7.7789ZM1.8258,7.7789C1.8258,7.7132 1.8453,7.6491 1.8817,7.5945C1.9182,7.5399 1.97,7.4973 2.0307,7.4722C2.0914,7.4471 2.1581,7.4405 2.2225,7.4533C2.2869,7.4661 2.3461,7.4977 2.3925,7.5442C2.4389,7.5906 2.4705,7.6497 2.4833,7.7141C2.4961,7.7785 2.4896,7.8453 2.4644,7.9059C2.4393,7.9666 2.3968,8.0184 2.3422,8.0549C2.2876,8.0914 2.2234,8.1109 2.1577,8.1109C2.0754,8.1063 1.9977,8.0714 1.9398,8.0127C1.8818,7.9541 1.8478,7.876 1.8442,7.7937L1.8258,7.7789Z"
+      android:fillColor="#686868"/>
+  <path
+      android:pathData="M2.7774,11.1427C2.7774,11.0202 2.7411,10.9004 2.673,10.7985C2.6049,10.6966 2.5081,10.6171 2.3949,10.5702C2.2817,10.5233 2.1571,10.5111 2.0369,10.535C1.9167,10.5589 1.8062,10.6179 1.7196,10.7046C1.6329,10.7912 1.5739,10.9016 1.55,11.0218C1.5261,11.142 1.5384,11.2666 1.5853,11.3799C1.6322,11.4931 1.7116,11.5899 1.8135,11.658C1.9154,11.726 2.0352,11.7624 2.1577,11.7624C2.3221,11.7624 2.4797,11.6971 2.5959,11.5809C2.7121,11.4647 2.7774,11.3071 2.7774,11.1427ZM1.8258,11.1427C1.8258,11.0771 1.8453,11.0129 1.8817,10.9583C1.9182,10.9037 1.97,10.8612 2.0307,10.836C2.0914,10.8109 2.1581,10.8043 2.2225,10.8171C2.2869,10.83 2.3461,10.8616 2.3925,10.908C2.4389,10.9544 2.4705,11.0136 2.4833,11.078C2.4961,11.1424 2.4896,11.2091 2.4644,11.2698C2.4393,11.3304 2.3968,11.3823 2.3422,11.4187C2.2876,11.4552 2.2234,11.4747 2.1577,11.4747C2.0703,11.4699 1.9883,11.4308 1.9296,11.3659C1.8709,11.3009 1.8402,11.2154 1.8442,11.128L1.8258,11.1427Z"
+      android:fillColor="#686868"/>
+  <path
+      android:pathData="M2.7774,14.4661C2.7781,14.3433 2.7424,14.2232 2.6748,14.1208C2.6071,14.0184 2.5106,13.9384 2.3974,13.8909C2.2843,13.8435 2.1596,13.8307 2.0391,13.8542C1.9187,13.8777 1.8079,13.9364 1.7209,14.0229C1.6339,14.1094 1.5745,14.2198 1.5503,14.3401C1.5261,14.4604 1.5381,14.5852 1.5849,14.6986C1.6317,14.8121 1.7111,14.9091 1.8131,14.9773C1.9151,15.0456 2.035,15.082 2.1577,15.082C2.3211,15.0811 2.4776,15.016 2.5935,14.9008C2.7094,14.7856 2.7755,14.6295 2.7774,14.4661ZM1.8258,14.4329C1.8258,14.3448 1.8607,14.2604 1.923,14.1981C1.9852,14.1359 2.0697,14.1009 2.1577,14.1009C2.2458,14.1009 2.3302,14.1359 2.3925,14.1981C2.4547,14.2604 2.4897,14.3448 2.4897,14.4329C2.4897,14.5209 2.4547,14.6053 2.3925,14.6676C2.3302,14.7298 2.2458,14.7648 2.1577,14.7648C2.0697,14.7648 1.9852,14.7298 1.923,14.6676C1.8607,14.6053 1.8258,14.5209 1.8258,14.4329Z"
+      android:fillColor="#686868"/>
+</vector>
