+++ app/src/main/res/drawable/ic_klasifikasi.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="20dp"
+    android:height="18dp"
+    android:viewportWidth="20"
+    android:viewportHeight="18">
+  <path
+      android:pathData="M19.3687,14.874H18.5176V5.2438C18.5253,4.9983 18.4354,4.7597 18.2675,4.5805C18.0995,4.4012 17.8674,4.2958 17.6219,4.2875H14.8804C14.7623,4.293 14.6465,4.3225 14.5402,4.3741C14.4339,4.4258 14.3392,4.4987 14.2619,4.5881C14.1847,4.6776 14.1265,4.7819 14.0908,4.8947C14.0552,5.0074 14.0429,5.1262 14.0548,5.2438V14.874H12.3971C12.3971,14.6923 12.3971,14.6126 12.3971,14.4947V1.2113C12.3971,0.5005 11.9955,0 11.495,0H8.7535C8.2531,0 7.7685,0.5005 7.7685,1.2113L7.7494,14.8294V14.8772H6.4456V8.5655C6.4456,8.218 6.0312,7.9279 5.5307,7.9279H2.7893C2.2888,7.9279 1.817,8.2116 1.817,8.5655V14.8772H0.5706C0.2518,14.8772 0,14.9856 0,15.1641V16.7102C0,16.8919 0.255,17.0289 0.5706,17.0289H19.3624C19.6811,17.0289 19.9999,16.895 19.9999,16.7102V15.1641C20.0063,14.9824 19.6843,14.874 19.3687,14.874Z"
+      android:fillColor="#686868"/>
+</vector>
