+++ app/src/main/res/drawable/ic_lokasi_rujukan.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="33dp"
+    android:height="33dp"
+    android:viewportWidth="33"
+    android:viewportHeight="33">
+  <path
+      android:pathData="M16.5,16.5m-16.5,0a16.5,16.5 0,1 1,33 0a16.5,16.5 0,1 1,-33 0"
+      android:fillColor="#F43B3B"
+      android:fillAlpha="0.3"/>
+  <path
+      android:pathData="M26.9997,14.0008C26.9998,12.5658 26.6911,11.1477 26.0946,9.8426C25.498,8.5375 24.6276,7.3761 23.5424,6.4373C22.4572,5.4985 21.1826,4.8042 19.8053,4.4016C18.428,3.999 16.9802,3.8976 15.5602,4.1042C14.1402,4.3108 12.7813,4.8206 11.5757,5.5989C10.3702,6.3772 9.3463,7.4059 8.5736,8.6151C7.8009,9.8242 7.2975,11.1855 7.0976,12.6064C6.8976,14.0273 7.0058,15.4747 7.4148,16.8501C8.837,22.6081 17.0443,29.0475 17.0443,29.0475C17.0443,29.0475 26.3182,21.7735 26.8762,15.7835H26.8516C26.9548,15.1949 27.0043,14.5983 26.9997,14.0008ZM17.1233,16.2279C16.3132,16.2289 15.521,15.9895 14.8471,15.54C14.1731,15.0906 13.6476,14.4513 13.3372,13.703C13.0267,12.9548 12.9452,12.1313 13.103,11.3367C13.2608,10.5421 13.6508,9.8122 14.2236,9.2394C14.7965,8.6666 15.5264,8.2766 16.3209,8.1188C17.1155,7.961 17.939,8.0425 18.6873,8.353C19.4355,8.6634 20.0748,9.1889 20.5243,9.8628C20.9737,10.5368 21.2131,11.329 21.2121,12.1391C21.2108,13.2205 20.7817,14.2576 20.0183,15.0237C19.255,15.7898 18.2196,16.2227 17.1381,16.2279H17.1233Z"
+      android:fillColor="#F43B3B"/>
+</vector>
