+++ app/src/main/res/drawable/ic_mengantisipasi.xml
@@ -1,0 +1,32 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="43dp"
+    android:height="43dp"
+    android:viewportWidth="43"
+    android:viewportHeight="43">
+  <path
+      android:pathData="M21.5,21.5m-21.5,0a21.5,21.5 0,1 1,43 0a21.5,21.5 0,1 1,-43 0"
+      android:fillColor="#00B185"
+      android:fillAlpha="0.3"/>
+  <group>
+    <clip-path
+        android:pathData="M21.5,21.5m-21.5,0a21.5,21.5 0,1 1,43 0a21.5,21.5 0,1 1,-43 0"/>
+    <path
+        android:pathData="M35.4738,29.3495C35.4739,29.0486 35.5589,28.7538 35.7189,28.4989L36.883,26.6609C37.0422,26.4057 37.127,26.1111 37.1281,25.8103V13.0342C37.1339,12.7385 37.041,12.4494 36.8641,12.2125C36.6872,11.9756 36.4365,11.8044 36.1514,11.7259C35.9582,11.6819 35.7575,11.6827 35.5647,11.728C35.3718,11.7734 35.1918,11.8622 35.0385,11.9877C34.8852,12.1132 34.7626,12.2721 34.68,12.4522C34.5974,12.6323 34.5571,12.8289 34.562,13.0269V17.2833C34.562,17.3483 34.5362,17.4106 34.4902,17.4566C34.4443,17.5025 34.3819,17.5283 34.3169,17.5283C34.2519,17.5283 34.1896,17.5025 34.1436,17.4566C34.0977,17.4106 34.0719,17.3483 34.0719,17.2833V10.2951C34.0793,9.9684 33.9662,9.6504 33.7542,9.4018C33.5421,9.1532 33.2459,8.9914 32.9222,8.9472C32.7435,8.9285 32.5628,8.9476 32.3919,9.0032C32.221,9.0588 32.0637,9.1497 31.9301,9.27C31.7966,9.3903 31.6898,9.5372 31.6167,9.7014C31.5436,9.8656 31.5059,10.0433 31.5058,10.223V16.5409C31.5049,16.6052 31.4786,16.6666 31.4328,16.7118C31.3869,16.757 31.3251,16.7823 31.2608,16.7823C31.1967,16.7823 31.1353,16.7569 31.09,16.7116C31.0447,16.6663 31.0193,16.6049 31.0193,16.5409V9.3581C31.0259,9.0306 30.9121,8.7122 30.6996,8.4631C30.487,8.214 30.1904,8.0515 29.866,8.0066C29.6871,7.9884 29.5063,8.008 29.3354,8.064C29.1645,8.1201 29.0073,8.2113 28.8738,8.3319C28.7403,8.4524 28.6336,8.5996 28.5606,8.764C28.4875,8.9283 28.4497,9.1061 28.4496,9.286V17.2833C28.4538,17.3177 28.4507,17.3526 28.4404,17.3858C28.4301,17.4189 28.4128,17.4494 28.3898,17.4754C28.3668,17.5014 28.3386,17.5222 28.307,17.5364C28.2753,17.5506 28.241,17.558 28.2064,17.558C28.1717,17.558 28.1374,17.5506 28.1057,17.5364C28.0741,17.5222 28.0459,17.5014 28.0229,17.4754C27.9999,17.4494 27.9827,17.4189 27.9723,17.3858C27.962,17.3526 27.9589,17.3177 27.9631,17.2833V11.6502C27.9704,11.323 27.8568,11.0046 27.644,10.7559C27.4312,10.5072 27.1342,10.3457 26.8098,10.3023C26.6311,10.2836 26.4504,10.3027 26.2795,10.3583C26.1086,10.4139 25.9513,10.5048 25.8178,10.6251C25.6842,10.7454 25.5774,10.8924 25.5043,11.0565C25.4313,11.2207 25.3935,11.3984 25.3934,11.5781V21.6189C25.3934,21.747 25.3425,21.8698 25.252,21.9604C25.1614,22.0509 25.0386,22.1018 24.9105,22.1018V20.0944C24.9095,19.4135 24.6384,18.7608 24.1566,18.2797C23.6748,17.7986 23.0217,17.5283 22.3408,17.5283V23.8786C22.341,24.1743 22.4129,24.4655 22.5504,24.7273C22.6878,24.989 22.8868,25.2135 23.1301,25.3815L25.7106,27.1618C25.9539,27.329 26.1528,27.5529 26.2903,27.814C26.4278,28.0752 26.4997,28.3659 26.4999,28.6611V30.8524H35.4738V29.3495Z"
+        android:fillColor="#00B185"/>
+    <path
+        android:pathData="M35.4738,30.8524H26.4999V31.7137H35.4738V30.8524Z"
+        android:fillColor="#579985"/>
+    <path
+        android:pathData="M35.4738,31.7137H26.4999V46H35.4738V31.7137Z"
+        android:fillColor="#F0C6AD"/>
+    <path
+        android:pathData="M7.6545,29.3495C7.6544,29.0486 7.5695,28.7538 7.4094,28.4989L6.2453,26.6609C6.0862,26.4057 6.0013,26.1111 6.0003,25.8103V13.0342C5.9945,12.7385 6.0873,12.4494 6.2642,12.2125C6.4411,11.9756 6.6919,11.8044 6.9769,11.7259C7.1701,11.6819 7.3708,11.6827 7.5637,11.728C7.7565,11.7734 7.9365,11.8622 8.0898,11.9877C8.2431,12.1132 8.3657,12.2721 8.4483,12.4522C8.5309,12.6323 8.5712,12.8289 8.5663,13.0269V17.2833C8.5621,17.3177 8.5653,17.3526 8.5756,17.3858C8.5859,17.4189 8.6031,17.4494 8.6261,17.4754C8.6491,17.5014 8.6773,17.5222 8.709,17.5364C8.7406,17.5507 8.7749,17.558 8.8096,17.558C8.8443,17.558 8.8786,17.5507 8.9102,17.5364C8.9418,17.5222 8.9701,17.5014 8.9931,17.4754C9.0161,17.4494 9.0333,17.4189 9.0436,17.3858C9.0539,17.3526 9.057,17.3177 9.0528,17.2833V10.2951C9.0463,9.9681 9.1602,9.6501 9.3729,9.4016C9.5855,9.1531 9.882,8.9913 10.2061,8.9472C10.3849,8.9285 10.5655,8.9476 10.7364,9.0032C10.9073,9.0588 11.0646,9.1497 11.1982,9.27C11.3317,9.3903 11.4385,9.5372 11.5116,9.7014C11.5847,9.8656 11.6225,10.0433 11.6225,10.223V16.5409C11.6297,16.6003 11.6585,16.655 11.7033,16.6947C11.7481,16.7344 11.8059,16.7563 11.8658,16.7563C11.9256,16.7563 11.9834,16.7344 12.0282,16.6947C12.0731,16.655 12.1018,16.6003 12.109,16.5409V9.3581C12.1024,9.0306 12.2162,8.7122 12.4288,8.4631C12.6413,8.214 12.9379,8.0515 13.2623,8.0066C13.4413,7.9884 13.622,8.008 13.7929,8.064C13.9638,8.1201 14.1211,8.2113 14.2545,8.3319C14.388,8.4524 14.4947,8.5996 14.5678,8.764C14.6408,8.9283 14.6786,9.1061 14.6787,9.286V17.2833C14.6745,17.3177 14.6777,17.3526 14.688,17.3858C14.6983,17.4189 14.7155,17.4494 14.7385,17.4754C14.7615,17.5014 14.7897,17.5222 14.8214,17.5364C14.853,17.5507 14.8873,17.558 14.922,17.558C14.9567,17.558 14.991,17.5507 15.0226,17.5364C15.0542,17.5222 15.0825,17.5014 15.1055,17.4754C15.1285,17.4494 15.1457,17.4189 15.156,17.3858C15.1663,17.3526 15.1694,17.3177 15.1652,17.2833V11.6502C15.1579,11.323 15.2715,11.0046 15.4843,10.7559C15.6971,10.5072 15.9941,10.3457 16.3185,10.3023C16.4971,10.2836 16.6776,10.3027 16.8483,10.3583C17.019,10.4139 17.1761,10.5049 17.3094,10.6252C17.4426,10.7456 17.5491,10.8926 17.6218,11.0568C17.6944,11.221 17.7318,11.3986 17.7313,11.5781V21.6189C17.7322,21.7473 17.7839,21.8701 17.8751,21.9606C17.9662,22.0511 18.0894,22.1018 18.2178,22.1018V20.0944C18.2188,19.4135 18.4899,18.7608 18.9717,18.2797C19.4535,17.7986 20.1066,17.5283 20.7875,17.5283V23.8786C20.7873,24.1743 20.7155,24.4655 20.578,24.7273C20.4405,24.989 20.2416,25.2135 19.9982,25.3815L17.4177,27.1618C17.1745,27.329 16.9755,27.5529 16.838,27.814C16.7005,28.0752 16.6286,28.3659 16.6285,28.6611V30.8524H7.6545V29.3495Z"
+        android:fillColor="#00B185"/>
+    <path
+        android:pathData="M16.6285,30.8524H7.6545V31.7137H16.6285V30.8524Z"
+        android:fillColor="#579985"/>
+    <path
+        android:pathData="M16.6285,31.7137H7.6545V46H16.6285V31.7137Z"
+        android:fillColor="#F0C6AD"/>
+  </group>
+</vector>
