+++ app/src/main/res/drawable/ic_rujukan.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="20dp"
+    android:height="19dp"
+    android:viewportWidth="20"
+    android:viewportHeight="19">
+  <path
+      android:pathData="M18.1735,4.4367H16.1302V0H3.8838V4.4367H1.8405C1.5989,4.4362 1.3597,4.4834 1.1364,4.5755C0.9131,4.6676 0.7102,4.8028 0.5393,4.9735C0.3683,5.1441 0.2327,5.3467 0.1402,5.5698C0.0476,5.7929 -0,6.0321 0,6.2736V16.3891C0,16.8772 0.1939,17.3453 0.5391,17.6905C0.8842,18.0356 1.3523,18.2295 1.8405,18.2295H18.1595C18.6477,18.2295 19.1158,18.0356 19.4609,17.6905C19.8061,17.3453 20,16.8772 20,16.3891V6.2736C20,5.7882 19.8079,5.3226 19.4657,4.9784C19.1235,4.6342 18.6589,4.4394 18.1735,4.4367ZM5.4654,1.5605H14.5626V4.4367H5.4654V1.5605ZM18.1246,11.9069H15.5948V14.4367H14.4472V11.9069H11.9174V10.7558H14.4472V8.24H15.5948V10.7663H18.1246V11.9069Z"
+      android:fillColor="#686868"/>
+</vector>
