+++ app/src/main/res/layout/fragment_home.xml
@@ -45,6 +45,7 @@
     <androidx.core.widget.NestedScrollView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
+        android:layout_marginBottom="56dp"
         app:layout_behavior="@string/appbar_scrolling_view_behavior">
 
         <androidx.constraintlayout.widget.ConstraintLayout
@@ -111,11 +112,13 @@
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintTop_toBottomOf="@+id/locationTextInputLayout" />
 
-            <RelativeLayout
+            <androidx.constraintlayout.widget.ConstraintLayout
                 android:id="@+id/relativeLayout"
-                android:layout_width="wrap_content"
+                android:layout_width="match_parent"
                 android:layout_height="90dp"
                 android:layout_marginTop="8dp"
+                android:layout_marginStart="10dp"
+                android:layout_marginEnd="10dp"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toBottomOf="@+id/textHeading1">
@@ -125,32 +128,59 @@
                     layout="@layout/shape_terkonfirmasi"
                     android:layout_width="80dp"
                     android:layout_height="90dp"
-                    android:layout_marginEnd="8dp" />
+                    app:layout_constraintStart_toStartOf="parent" />
 
                 <include
                     android:id="@+id/include2"
                     layout="@layout/shape_kasus_aktif"
                     android:layout_width="80dp"
                     android:layout_height="90dp"
-                    android:layout_marginEnd="8dp"
-                    android:layout_toEndOf="@id/include" />
+                    android:layout_marginStart="8dp"
+                    app:layout_constraintStart_toEndOf="@+id/include" />
 
                 <include
                     android:id="@+id/include3"
                     layout="@layout/shape_sembuh"
                     android:layout_width="80dp"
                     android:layout_height="90dp"
-                    android:layout_marginEnd="8dp"
-                    android:layout_toEndOf="@id/include2" />
+                    android:layout_marginStart="8dp"
+                    app:layout_constraintStart_toEndOf="@+id/include2" />
 
                 <include
                     android:id="@+id/include4"
                     layout="@layout/shape_meninggal"
                     android:layout_width="80dp"
                     android:layout_height="90dp"
-                    android:layout_toEndOf="@id/include3" />
+                    android:layout_marginStart="8dp"
+                    app:layout_constraintStart_toEndOf="@+id/include3" />
 
-            </RelativeLayout>
+<!--                <TextView-->
+<!--                    android:id="@+id/textView5"-->
+<!--                    android:layout_width="0.1dp"-->
+<!--                    android:layout_height="match_parent"-->
+<!--                    android:minWidth="0dp"-->
+<!--                    android:layout_marginEnd="8dp"-->
+<!--                    app:layout_constraintBottom_toBottomOf="parent"-->
+<!--                    app:layout_constraintEnd_toEndOf="parent"-->
+<!--                    app:layout_constraintStart_toStartOf="parent"-->
+<!--                    app:layout_constraintTop_toTopOf="parent" />-->
+
+
+<!--                <androidx.constraintlayout.widget.Guideline-->
+<!--                    android:id="@+id/guideline"-->
+<!--                    android:layout_width="wrap_content"-->
+<!--                    android:layout_height="match_parent"-->
+<!--                    android:orientation="vertical"-->
+<!--                    app:layout_constraintGuide_end="184dp" />-->
+
+<!--                <androidx.constraintlayout.widget.Guideline-->
+<!--                    android:id="@+id/guideline2"-->
+<!--                    android:layout_width="wrap_content"-->
+<!--                    android:layout_height="match_parent"-->
+<!--                    android:orientation="vertical"-->
+<!--                    app:layout_constraintGuide_begin="184dp" />-->
+
+            </androidx.constraintlayout.widget.ConstraintLayout>
 
             <TextView
                 android:id="@+id/textDateJudul"
@@ -188,9 +218,11 @@
             <include
                 android:id="@+id/include5"
                 layout="@layout/shape_peta_sebaran"
-                android:layout_width="344dp"
+                android:layout_width="match_parent"
                 android:layout_height="180dp"
                 android:layout_marginTop="8dp"
+                android:layout_marginStart="10dp"
+                android:layout_marginEnd="10dp"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toBottomOf="@+id/textHeading2" />
@@ -219,5 +251,15 @@
         </androidx.constraintlayout.widget.ConstraintLayout>
 
     </androidx.core.widget.NestedScrollView>
+
+    <com.google.android.material.bottomnavigation.BottomNavigationView
+        style="@style/BottomNavigationView"
+        android:id="@+id/bottom_navigation_id"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:background="@drawable/bg_bottom_navigation"
+        app:menu="@menu/bottom_navigation_menu"
+        app:labelVisibilityMode="labeled"
+        android:layout_gravity="bottom"/>
 
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
