+++ app/src/main/res/layout/item_row_news.xml
@@ -2,7 +2,7 @@
 <androidx.constraintlayout.widget.ConstraintLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="344dp"
+    android:layout_width="match_parent"
     android:layout_height="94dp"
     android:layout_marginBottom="8dp"
     android:background="@drawable/bg_shape">
