+++ app/src/main/res/layout/shape_peta_sebaran.xml
@@ -1,10 +1,29 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="344dp"
     android:layout_height="180dp"
     android:background="@drawable/bg_shape">
 
+<!--    <com.here.android.mpa.mapping.MapView-->
+<!--        android:id="@+id/mapViewConnection"-->
+<!--        android:layout_width="327dp"-->
+<!--        android:layout_height="166dp"-->
+<!--        app:layout_constraintBottom_toBottomOf="parent"-->
+<!--        app:layout_constraintEnd_toEndOf="parent"-->
+<!--        app:layout_constraintStart_toStartOf="parent"-->
+<!--        app:layout_constraintTop_toTopOf="parent" />-->
 
+<!--    <fragment-->
+<!--        android:id="@+id/mapDistribution"-->
+<!--        class="com.here.android.mpa.mapping.AndroidXMapFragment"-->
+<!--        android:layout_width="match_parent"-->
+<!--        android:layout_height="match_parent"-->
+<!--        android:layout_margin="8dp"-->
+<!--        app:layout_constraintBottom_toBottomOf="parent"-->
+<!--        app:layout_constraintEnd_toEndOf="parent"-->
+<!--        app:layout_constraintStart_toStartOf="parent"-->
+<!--        app:layout_constraintTop_toTopOf="parent" />-->
 
 </androidx.constraintlayout.widget.ConstraintLayout>
