+++ app/src/main/res/navigation/main_navigation.xml
@@ -18,4 +18,9 @@
         android:name="com.goggxi.covid19detector.ui.home.HomeFragment"
         android:label="fragment_home"
         tools:layout="@layout/fragment_home" />
+    <fragment
+        android:id="@+id/informasiFragment"
+        android:name="com.goggxi.covid19detector.ui.informasi.InformasiFragment"
+        android:label="fragment_informasi"
+        tools:layout="@layout/fragment_informasi" />
 </navigation>
